Add CMake-based pip install and system install
