+++ include/permanent.h
@@ -1,0 +1,8 @@
+/* Copyright 2024 QC-Devs (GPLv3) */
+
+#if !defined(permanent_h_)
+#define permanent_h_
+
+#include <permanent/permanent.h>
+
+#endif  // permanent_h_
