+++ include/permanent/tables.h
@@ -1,7 +1,7 @@
 /* Copyright 2024 QC-Devs (GPLv3) */
 
-#ifndef TABLES_H_
-#define TABLES_H_
+#if !defined(permanent_tables_h_)
+#define permanent_tables_h_
 
 #include <cstdint>
 
@@ -4240,28 +4240,23 @@ static const std::size_t binom_table[] = {
 };
 
 static const double factorial_table[] = {
-    1.0000000000000000e+00, 1.0000000000000000e+00, 2.0000000000000000e+00,
-    6.0000000000000000e+00, 2.4000000000000000e+01, 1.2000000000000000e+02,
-    7.2000000000000000e+02, 5.0400000000000000e+03, 4.0320000000000000e+04,
-    3.6288000000000000e+05, 3.6288000000000000e+06, 3.9916800000000000e+07,
-    4.7900160000000000e+08, 6.2270208000000000e+09, 8.7178291200000000e+10,
-    1.3076743680000000e+12, 2.0922789888000000e+13, 3.5568742809600000e+14,
-    6.4023737057280000e+15, 1.2164510040883200e+17, 2.4329020081766400e+18,
-    5.1090942171709440e+19, 1.1240007277776077e+21, 2.5852016738884978e+22,
-    6.2044840173323941e+23, 1.5511210043330984e+25, 4.0329146112660572e+26,
-    1.0888869450418352e+28, 3.0488834461171380e+29, 8.8417619937397008e+30,
-    2.6525285981219110e+32, 8.2228386541779236e+33, 2.6313083693369355e+35,
-    8.6833176188118871e+36, 2.9523279903960420e+38, 1.0333147966386147e+40,
-    3.7199332678990118e+41, 1.3763753091226346e+43, 5.2302261746660112e+44,
-    2.0397882081197442e+46, 8.1591528324789801e+47, 3.3452526613163818e+49,
-    1.4050061177528803e+51, 6.0415263063373834e+52, 2.6582715747884485e+54,
-    1.1962222086548019e+56, 5.5026221598120899e+57, 2.5862324151116818e+59,
-    1.2413915592536073e+61, 6.0828186403426752e+62, 3.0414093201713381e+64,
-    1.5511187532873824e+66, 8.0658175170943877e+67, 4.2748832840600255e+69,
-    2.3084369733924138e+71, 1.2696403353658278e+73, 7.1099858780486358e+74,
-    4.0526919504877214e+76, 2.3505613312828785e+78, 1.3868311854568984e+80,
-    8.3209871127413899e+81, 5.0758021387722484e+83, 3.1469973260387939e+85,
-    1.9826083154044403e+87, 1.2688693218588415e+89,
+    1.0000000000000000e+00, 1.0000000000000000e+00, 2.0000000000000000e+00, 6.0000000000000000e+00,
+    2.4000000000000000e+01, 1.2000000000000000e+02, 7.2000000000000000e+02, 5.0400000000000000e+03,
+    4.0320000000000000e+04, 3.6288000000000000e+05, 3.6288000000000000e+06, 3.9916800000000000e+07,
+    4.7900160000000000e+08, 6.2270208000000000e+09, 8.7178291200000000e+10, 1.3076743680000000e+12,
+    2.0922789888000000e+13, 3.5568742809600000e+14, 6.4023737057280000e+15, 1.2164510040883200e+17,
+    2.4329020081766400e+18, 5.1090942171709440e+19, 1.1240007277776077e+21, 2.5852016738884978e+22,
+    6.2044840173323941e+23, 1.5511210043330984e+25, 4.0329146112660572e+26, 1.0888869450418352e+28,
+    3.0488834461171380e+29, 8.8417619937397008e+30, 2.6525285981219110e+32, 8.2228386541779236e+33,
+    2.6313083693369355e+35, 8.6833176188118871e+36, 2.9523279903960420e+38, 1.0333147966386147e+40,
+    3.7199332678990118e+41, 1.3763753091226346e+43, 5.2302261746660112e+44, 2.0397882081197442e+46,
+    8.1591528324789801e+47, 3.3452526613163818e+49, 1.4050061177528803e+51, 6.0415263063373834e+52,
+    2.6582715747884485e+54, 1.1962222086548019e+56, 5.5026221598120899e+57, 2.5862324151116818e+59,
+    1.2413915592536073e+61, 6.0828186403426752e+62, 3.0414093201713381e+64, 1.5511187532873824e+66,
+    8.0658175170943877e+67, 4.2748832840600255e+69, 2.3084369733924138e+71, 1.2696403353658278e+73,
+    7.1099858780486358e+74, 4.0526919504877214e+76, 2.3505613312828785e+78, 1.3868311854568984e+80,
+    8.3209871127413899e+81, 5.0758021387722484e+83, 3.1469973260387939e+85, 1.9826083154044403e+87,
+    1.2688693218588415e+89,
 };
 
 static const std::uint64_t int_factorial_table[] = {
@@ -4288,4 +4283,4 @@ static const std::uint64_t int_factorial_table[] = {
     2432902008176640000,
 };
 
-#endif  // TABLES_H_
+#endif  // permanent_tables_h_
