+++ include/permanent/tuning.default.h
@@ -1,0 +1,15 @@
+/* Copyright 2024 QC-Devs (GPLv3) */
+
+#if !defined(permanent_tuning_h_)
+#define permanent_tuning_h_
+
+namespace permanent {
+
+constexpr double PARAM_1 = -0.572098;
+constexpr double PARAM_2 = -22.014212;
+constexpr double PARAM_3 = 15.29794;
+constexpr double PARAM_4 = 3.0;
+
+}  // namespace permanent
+
+#endif  // permanent_tuning_h_
