--- permanent/__init__.py
@@ -1,11 +1,0 @@
-r"""Module for computing the permanents of matrices."""
-
-__all__ = [
-    "opt",
-    "combinatoric",
-    "glynn",
-    "ryser",
-]
-
-
-from .permanent import opt, combinatoric, glynn, ryser
