+++ src/tuning.cc
@@ -1,5 +1,9 @@
 /* Copyright 2024 QC-Devs (GPLv3) */
 
+#define _PERMANENT_DEFAULT_TUNING
+
+#include <permanent.h>
+
 #include <cmath>
 #include <cstdlib>
 #include <ctime>
@@ -7,10 +11,6 @@
 #include <iomanip>
 #include <iostream>
 #include <string>
-
-#include "permanent.h"
-
-#ifdef RUN_TUNING
 
 constexpr char CSV_FILE[] = "src/tuning.csv";
 
@@ -26,256 +26,180 @@ constexpr std::size_t DATA_POINTS = MAX_ROWS * MAX_COLS;
 
 constexpr double TOLERANCE = 0.0001;
 
-#else
+int main(int argc, const char *argv[])
+{
+  /* Open CSV file */
 
-constexpr char HEADER_FILE[] = "src/tuning.h";
+  std::ofstream csv_file;
 
-constexpr double DEFAULT_PARAM_1 = -0.572098;
+  if (argc != 2) {
+    std::cerr << "No arguments passed" << std::endl;
+    return 1;
+  }
+  csv_file.open(argv[1]);
+  if (csv_file.fail()) {
+    std::cerr << "Cannot open CSV file" << std::endl;
+    return 2;
+  }
 
-constexpr double DEFAULT_PARAM_2 = -22.014212;
+  /* Set format options for printing */
 
-constexpr double DEFAULT_PARAM_3 = 15.297940;
+  std::cout.precision(9);
+  std::cerr.precision(9);
+  csv_file.precision(9);
 
-constexpr double DEFAULT_PARAM_4 = 3.0;
+  std::cout.fill(' ');
+  std::cerr.fill(' ');
+  csv_file.fill(' ');
 
-#endif  // RUN_TUNING
+  /* Print CSV headers */
 
-#ifdef RUN_TUNING
+  std::cout << CSV_HEADER << std::endl;
 
-int main(int argc, char *argv[]) {
-    (void)argc;
-    (void)argv;
+  csv_file << CSV_HEADER << '\n';
 
-    /* Open CSV file */
+  if (csv_file.fail()) {
+    std::cerr << "Error writing to CSV file" << std::endl;
+    csv_file.close();
+    return 2;
+  }
 
-    std::ofstream csv_file;
+  /* Time efficiency of algorithms for different size matrices */
 
-    csv_file.open(CSV_FILE);
+  std::size_t i, m, n;
 
-    if (csv_file.fail()) {
-        std::cerr << "Cannot open CSV file" << std::endl;
-        return 2;
-    }
+  std::clock_t begin, end;
 
-    /* Set format options for printing */
+  double soln_combn, soln_glynn, soln_ryser;
+  double mean_combn, mean_glynn, mean_ryser;
 
-    std::cout.precision(9);
-    std::cerr.precision(9);
-    csv_file.precision(9);
+  double time_combn[128];
+  double time_glynn[128];
+  double time_ryser[128];
 
-    std::cout.fill(' ');
-    std::cerr.fill(' ');
-    csv_file.fill(' ');
+  double array[MAX_ROWS * MAX_COLS];
 
-    /* Print CSV headers */
+  int fastest;
 
-    std::cout << CSV_HEADER << std::endl;
+  /* Random binary matrix for testing. */
 
-    csv_file << CSV_HEADER << '\n';
+  std::srand(1234789789U);
+  for (i = 0; i < MAX_ROWS * MAX_COLS; ++i) {
+    array[i] = (std::rand() / static_cast<double>(RAND_MAX) - 0.5) * 2;
+  }
 
-    if (csv_file.fail()) {
+  /* Iterate over number of rows and number of columns. */
+
+  for (m = 2; m <= MAX_ROWS; ++m) {
+    for (n = m; n <= MAX_COLS; ++n) {
+      /* Solve the permanent using each algorithm NUM_TRIALS number of
+       * times. */
+
+      if (m == n) {
+        for (i = 0; i != NUM_TRIALS; ++i) {
+          begin = std::clock();
+          soln_combn = permanent::combinatoric<double>(m, n, array);
+          end = std::clock();
+          time_combn[i] = static_cast<double>(end - begin);
+
+          begin = std::clock();
+          soln_glynn = permanent::glynn<double>(m, n, array);
+          end = std::clock();
+          time_glynn[i] = static_cast<double>(end - begin);
+
+          begin = std::clock();
+          soln_ryser = permanent::ryser<double>(m, n, array);
+          end = std::clock();
+          time_ryser[i] = static_cast<double>(end - begin);
+
+          if ((std::fabs(soln_combn - soln_glynn) / soln_combn > TOLERANCE) ||
+              (std::fabs(soln_combn - soln_ryser) / soln_ryser > TOLERANCE)) {
+            std::cerr << "Bad permanent values:"
+                      << "\nCombn: " << soln_combn << "\nGlynn: " << soln_glynn
+                      << "\nRyser: " << soln_ryser << std::endl;
+            csv_file.close();
+            return 1;
+          }
+        }
+      } else {
+        for (i = 0; i != NUM_TRIALS; ++i) {
+          begin = std::clock();
+          soln_combn = permanent::combinatoric_rectangular<double>(m, n, array);
+          end = std::clock();
+          time_combn[i] = static_cast<double>(end - begin);
+
+          begin = std::clock();
+          soln_glynn = permanent::glynn_rectangular<double>(m, n, array);
+          end = std::clock();
+          time_glynn[i] = static_cast<double>(end - begin);
+
+          begin = std::clock();
+          soln_ryser = permanent::ryser_rectangular<double>(m, n, array);
+          end = std::clock();
+          time_ryser[i] = static_cast<double>(end - begin);
+
+          if ((std::fabs(soln_combn - soln_glynn) / soln_combn > TOLERANCE) ||
+              (std::fabs(soln_combn - soln_ryser) / soln_ryser > TOLERANCE)) {
+            std::cerr << std::scientific << "Bad permanent values:"
+                      << "\nCombn: " << soln_combn << "\nGlynn: " << soln_glynn
+                      << "\nRyser: " << soln_ryser << std::endl;
+            csv_file.close();
+            return 1;
+          }
+        }
+      }
+
+      /* Calculate the mean for the runtime of each algorithm. */
+
+      mean_combn = 0.0;
+      mean_glynn = 0.0;
+      mean_ryser = 0.0;
+
+      for (i = 0; i != NUM_TRIALS; ++i) {
+        mean_combn += time_combn[i];
+        mean_glynn += time_glynn[i];
+        mean_ryser += time_ryser[i];
+      }
+
+      mean_combn = mean_combn / NUM_TRIALS;
+      mean_glynn = mean_glynn / NUM_TRIALS;
+      mean_ryser = mean_ryser / NUM_TRIALS;
+
+      /* Find the fastest algorithm */
+
+      if (mean_ryser <= mean_glynn) {
+        fastest = 0;
+      } else if (mean_combn <= mean_glynn) {
+        fastest = 1;
+      } else {
+        fastest = 2;
+      }
+
+      /* Write line */
+
+      std::cout << std::scientific << static_cast<double>(m) / n << ',' << std::setw(2) << n << ','
+                << std::scientific << mean_combn << ',' << mean_glynn << ',' << std::scientific
+                << mean_ryser << ',' << fastest << std::endl;
+
+      csv_file << std::scientific << static_cast<double>(m) / n << ',' << std::setw(2) << n << ','
+               << std::scientific << mean_combn << ',' << mean_glynn << ',' << std::scientific
+               << mean_ryser << ',' << fastest << '\n';
+
+      if (csv_file.fail()) {
         std::cerr << "Error writing to CSV file" << std::endl;
         csv_file.close();
         return 2;
+      }
     }
+  }
 
-    /* Time efficiency of algorithms for different size matrices */
+  /* Close CSV file */
 
-    std::size_t i, m, n;
+  csv_file.close();
 
-    std::clock_t begin, end;
+  /* Exit successfully */
 
-    double soln_combn, soln_glynn, soln_ryser;
-    double mean_combn, mean_glynn, mean_ryser;
-
-    double time_combn[128];
-    double time_glynn[128];
-    double time_ryser[128];
-
-    double array[MAX_ROWS * MAX_COLS];
-
-    int fastest;
-
-    /* Random binary matrix for testing. */
-
-    std::srand(1234789789U);
-    for (i = 0; i < MAX_ROWS * MAX_COLS; ++i) {
-        array[i] = (std::rand() / static_cast<double>(RAND_MAX) - 0.5) * 2;
-    }
-
-    /* Iterate over number of rows and number of columns. */
-
-    for (m = 2; m <= MAX_ROWS; ++m) {
-        for (n = m; n <= MAX_COLS; ++n) {
-            /* Solve the permanent using each algorithm NUM_TRIALS number of
-             * times. */
-
-            if (m == n) {
-                for (i = 0; i != NUM_TRIALS; ++i) {
-                    begin = std::clock();
-                    soln_combn = permanent::combinatoric<double>(m, n, array);
-                    end = std::clock();
-                    time_combn[i] = static_cast<double>(end - begin);
-
-                    begin = std::clock();
-                    soln_glynn = permanent::glynn<double>(m, n, array);
-                    end = std::clock();
-                    time_glynn[i] = static_cast<double>(end - begin);
-
-                    begin = std::clock();
-                    soln_ryser = permanent::ryser<double>(m, n, array);
-                    end = std::clock();
-                    time_ryser[i] = static_cast<double>(end - begin);
-
-                    if ((std::fabs(soln_combn - soln_glynn) / soln_combn >
-                         TOLERANCE) ||
-                        (std::fabs(soln_combn - soln_ryser) / soln_ryser >
-                         TOLERANCE)) {
-                        std::cerr << "Bad permanent values:"
-                                  << "\nCombn: " << soln_combn
-                                  << "\nGlynn: " << soln_glynn
-                                  << "\nRyser: " << soln_ryser << std::endl;
-                        csv_file.close();
-                        return 1;
-                    }
-                }
-            } else {
-                for (i = 0; i != NUM_TRIALS; ++i) {
-                    begin = std::clock();
-                    soln_combn = permanent::combinatoric_rectangular<double>(m, n, array);
-                    end = std::clock();
-                    time_combn[i] = static_cast<double>(end - begin);
-
-                    begin = std::clock();
-                    soln_glynn = permanent::glynn_rectangular<double>(m, n, array);
-                    end = std::clock();
-                    time_glynn[i] = static_cast<double>(end - begin);
-
-                    begin = std::clock();
-                    soln_ryser = permanent::ryser_rectangular<double>(m, n, array);
-                    end = std::clock();
-                    time_ryser[i] = static_cast<double>(end - begin);
-
-                    if ((std::fabs(soln_combn - soln_glynn) / soln_combn >
-                         TOLERANCE) ||
-                        (std::fabs(soln_combn - soln_ryser) / soln_ryser >
-                         TOLERANCE)) {
-                        std::cerr << std::scientific << "Bad permanent values:"
-                                  << "\nCombn: " << soln_combn
-                                  << "\nGlynn: " << soln_glynn
-                                  << "\nRyser: " << soln_ryser << std::endl;
-                        csv_file.close();
-                        return 1;
-                    }
-                }
-            }
-
-            /* Calculate the mean for the runtime of each algorithm. */
-
-            mean_combn = 0.0;
-            mean_glynn = 0.0;
-            mean_ryser = 0.0;
-
-            for (i = 0; i != NUM_TRIALS; ++i) {
-                mean_combn += time_combn[i];
-                mean_glynn += time_glynn[i];
-                mean_ryser += time_ryser[i];
-            }
-
-            mean_combn = mean_combn / NUM_TRIALS;
-            mean_glynn = mean_glynn / NUM_TRIALS;
-            mean_ryser = mean_ryser / NUM_TRIALS;
-
-            /* Find the fastest algorithm */
-
-            if (mean_ryser <= mean_glynn) {
-                fastest = 0;
-            } else if (mean_combn <= mean_glynn) {
-                fastest = 1;
-            } else {
-                fastest = 2;
-            }
-
-            /* Write line */
-
-            std::cout << std::scientific << static_cast<double>(m) / n << ','
-                      << std::setw(2) << n << ',' << std::scientific
-                      << mean_combn << ',' << mean_glynn << ','
-                      << std::scientific << mean_ryser << ',' << fastest
-                      << std::endl;
-
-            csv_file << std::scientific << static_cast<double>(m) / n << ','
-                     << std::setw(2) << n << ',' << std::scientific
-                     << mean_combn << ',' << mean_glynn << ','
-                     << std::scientific << mean_ryser << ',' << fastest << '\n';
-
-            if (csv_file.fail()) {
-                std::cerr << "Error writing to CSV file" << std::endl;
-                csv_file.close();
-                return 2;
-            }
-        }
-    }
-
-    /* Close CSV file */
-
-    csv_file.close();
-
-    /* Exit successfully */
-
-    return 0;
+  return 0;
 }
 
-#else
-
-int main(int argc, char *argv[]) {
-    (void)argc;
-    (void)argv;
-
-    /* Open header file */
-
-    std::ofstream header_file;
-
-    header_file.open(HEADER_FILE);
-
-    if (header_file.fail()) {
-        std::cerr << "Cannot open header file" << std::endl;
-        return 2;
-    }
-
-    /* Set format options for printing */
-
-    header_file.precision(9);
-
-    /* Write default header file */
-
-    header_file << "#ifndef PERMANENT_TUNING_H_\n";
-    header_file << "#define PERMANENT_TUNING_H_\n";
-    header_file << "\n\n";
-    header_file << "constexpr double PARAM_1 = " << std::scientific
-                << DEFAULT_PARAM_1 << ";\n";
-    header_file << "constexpr double PARAM_2 = " << std::scientific
-                << DEFAULT_PARAM_2 << ";\n";
-    header_file << "constexpr double PARAM_3 = " << std::scientific
-                << DEFAULT_PARAM_3 << ";\n";
-    header_file << "constexpr double PARAM_4 = " << std::scientific
-                << DEFAULT_PARAM_4 << ";\n";
-    header_file << "\n\n";
-    header_file << "#endif  // PERMANENT_TUNING_H_\n";
-
-    if (header_file.fail()) {
-        std::cerr << "Error writing to header file" << std::endl;
-        header_file.close();
-        return 2;
-    }
-
-    /* Close header file */
-
-    header_file.close();
-
-    /* Exit successfully */
-
-    return 0;
-}
-
-#endif  // RUN_TUNING
+#undef _PERMANENT_DEFAULT_TUNING
