--- tools/include_dirs.py
@@ -1,6 +1,0 @@
-import sysconfig
-
-import numpy
-
-
-print(f"-I{sysconfig.get_paths()['include']} -I{numpy.get_include()}")
